+++ FE/src/App.jsx
@@ -1,15 +1,13 @@
 import "./App.css";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Footer from "./components/Footer";
-import Header from "./components/Header";
 import Home from "./pages/Home";
 import ContextProvider from "./contexts/ContextProvider";
-import ProductDetail from "./components/ProductDetail";
 import Shop from "./pages/Shop";
 import Checkout from "./pages/Checkout";
 import OrderTracking from "./pages/OrderTracking";
 import Cart from "./pages/Cart";
 import Layout from "./layout/Layout";
+import ProductDetail from "./components/Product/ProductDetail";
 
 function App() {
   return (
