+++ FE/src/assets/style/cartItem.css
@@ -1,7 +1,9 @@
 .cart-product {
   border: 1px solid rgb(224, 220, 220);
   font-weight: 500;
+  height: fit-content;
 }
+
 .cart-product td {
   padding: 15px;
 }
@@ -25,9 +27,16 @@
   width: 30%;
   margin-right: 15px;
 }
+.product-info-flex {
+  display: flex;
+  align-items: center;
+}
 .product-name {
   width: 70%;
   color: #868686;
+}
+.product-stock {
+  display: inline-block;
 }
 .change-quantity-btn {
   max-width: 30%;
@@ -67,4 +76,15 @@
 .product-remove .remove-btn:hover {
   background-color: black;
   color: white;
+}
+
+.quantity::-webkit-outer-spin-button,
+.quantity::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+
+/* Firefox */
+.quantity[type="number"] {
+  -moz-appearance: textfield;
 }
