+++ FE/src/assets/style/cartPopup.css
@@ -7,6 +7,7 @@
   background-color: white;
   padding: 20px 15px;
   color: black;
+  z-index: 10;
 }
 .cart-popup-container a {
   color: black;
@@ -119,4 +120,15 @@
 /* Handle on hover */
 .cart-popup::-webkit-scrollbar-thumb:hover {
   background: #555;
+}
+
+.highlight-cart-popup {
+  position: fixed;
+  left: 0;
+  top: 0;
+  z-index: 10;
+  width: 100%;
+  height: 100%;
+  background-color: rgb(0, 0, 0);
+  background-color: rgba(0, 0, 0, 0.8);
 }
