+++ FE/src/assets/style/headerSlide.css
@@ -2,3 +2,6 @@
   width: 100%;
   height: 100vh;
 }
+.z-index-home button {
+  z-index: 0;
+}
