+++ FE/src/assets/style/subcribeForm.css
@@ -46,3 +46,6 @@
   color: red;
   margin-top: 10px;
 }
+.success-subcribe {
+  color: rgb(6, 196, 6);
+}
