+++ FE/src/components/Footer/index.jsx
@@ -1,8 +1,8 @@
-import "../assets/style/footer.css";
 import Brand from "./Brand";
 import SocialLink from "./SocialLink";
 import FooterShortLink from "./FooterShortLink";
 import TermAndPayment from "./TermAndPayment";
+import "../../assets/style/footer.css";
 
 const Footer = () => {
   return (
