+++ FE/src/components/Header/index.jsx
@@ -1,11 +1,11 @@
 import { useContext } from "react";
-import "../assets/style/header.css";
-import { Context } from "../contexts/ContextProvider";
+import "../../assets/style/header.css";
+import { ContextValue } from "../../contexts/ContextProvider";
 import UserMenu from "./UserMenu";
 import HeaderSlide from "./HeaderSlide";
 
 const Header = () => {
-  const { isHomePath } = useContext(Context);
+  const { isHomePath } = useContext(ContextValue);
   let checked;
   isHomePath.pathname === "/" ? (checked = true) : (checked = false);
   return (
