--- FE/src/components/HomeProducts.jsx
@@ -1,44 +1,0 @@
-import "../assets/style/homeProducts.css";
-import { useContext } from "react";
-import { Context } from "../contexts/ContextProvider";
-import ProductItem from "./ProductItem";
-const HomeProducts = () => {
-  const { showList, onShowHomeProducts, homeProductList } = useContext(Context);
-  return (
-    <section className="home-products">
-      <div className="show-change">
-        <button
-          className={`show-btn ${
-            showList === "Latest Products" ? "active-btn" : "deactive-btn"
-          }`}
-          onClick={() => onShowHomeProducts("Latest Products")}
-        >
-          <span className="highlight-show">Latest Products</span>
-        </button>
-        <button
-          className={`show-btn ${
-            showList === "Best Sellers" ? "active-btn" : "deactive-btn"
-          }`}
-          onClick={() => onShowHomeProducts("Best Sellers")}
-        >
-          <span className="highlight-show"> Best Sellers</span>
-        </button>
-        <button
-          className={`show-btn ${
-            showList === "Featured Products" ? "active-btn" : "deactive-btn"
-          }`}
-          onClick={() => onShowHomeProducts("Featured Products")}
-        >
-          <span className="highlight-show">Featured Products</span>
-        </button>
-      </div>
-      <div className="show-home-products">
-        {homeProductList.map((product) => (
-          <ProductItem key={product.id} product={product} />
-        ))}
-      </div>
-    </section>
-  );
-};
-
-export default HomeProducts;
