+++ FE/src/components/index.js
@@ -1,0 +1,5 @@
+import Header from "./Header";
+import Footer from "./Footer";
+import ProductList from "./Product/ProductList";
+import Cart from "../pages/Cart";
+export { Header, Footer, ProductList, Cart };
