+++ FE/src/layout/Layout.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import { Footer, Header } from "../components";
 
+//may kia tuong tu vay
 const Layout = () => {
   return (
     <>
