+++ FE/src/pages/Home.jsx
@@ -1,10 +1,10 @@
 import "../assets/style/home.css";
-import Category from "../components/Category";
-import HomeProducts from "../components/HomeProducts";
-import Banner from "../components/Banner";
-import ShopByStle from "../components/ShopByStle";
-import Collection from "../components/Collection";
-import SubcribeForm from "../components/SubcribeForm";
+import Category from "../components/Home/Category";
+import HomeProducts from "../components/Home/HomeProducts";
+import Banner from "../components/Home/Banner";
+import Collection from "../components/Home/Collection";
+import SubcribeForm from "../components/Home/SubcribeForm";
+import ShopByStyle from "../components/Home/ShopByStyle";
 const Home = () => {
   return (
     <section className="home">
@@ -12,7 +12,7 @@ const Home = () => {
       <Banner />
       <HomeProducts />
       <Collection />
-      <ShopByStle />
+      <ShopByStyle />
       <SubcribeForm />
     </section>
   );
