+++ FE/src/pages/Shop.jsx
@@ -1,4 +1,4 @@
-import ProductList from "../components/ProductList";
+import { ProductList } from "../components";
 
 const Shop = () => {
   return (
